# Use EffectFnN for some primitives