This compiles to better JS (no allocation of intermediate closures), and
gives us some performance boost:

Before:

```
dyn x 121,925 ops/sec ±0.64% (65 runs sampled)
dyn fmap x 119,804 ops/sec ±0.79% (62 runs sampled)
dyn ap pure x 93,008 ops/sec ±0.44% (66 runs sampled)
dyn ap self x 80,705 ops/sec ±0.68% (63 runs sampled)
dyn bind self x 20,436 ops/sec ±12.54% (29 runs sampled)
dyn bind inner x 81,020 ops/sec ±0.84% (62 runs sampled)
dyn bind outer x 59,404 ops/sec ±0.70% (61 runs sampled)
dyn 1ap  - fire first x 73,097 ops/sec ±0.97% (63 runs sampled)
dyn 5ap  - fire first x 36,428 ops/sec ±1.51% (64 runs sampled)
dyn 10ap - fire first x 23,141 ops/sec ±0.67% (66 runs sampled)
dyn 15ap - fire first x 16,916 ops/sec ±0.68% (65 runs sampled)
dyn 1ap  - fire last x 72,502 ops/sec ±0.90% (65 runs sampled)
dyn 5ap  - fire last x 35,133 ops/sec ±0.67% (65 runs sampled)
dyn 10ap - fire last x 22,049 ops/sec ±0.82% (62 runs sampled)
dyn 15ap - fire last x 16,501 ops/sec ±0.70% (62 runs sampled)
```

After:

```
dyn x 146,378 ops/sec ±0.75% (62 runs sampled)
dyn fmap x 140,482 ops/sec ±1.04% (62 runs sampled)
dyn ap pure x 120,909 ops/sec ±1.05% (63 runs sampled)
dyn ap self x 108,183 ops/sec ±0.98% (65 runs sampled)
dyn bind self x 21,429 ops/sec ±16.14% (24 runs sampled)
dyn bind inner x 112,964 ops/sec ±1.16% (64 runs sampled)
dyn bind outer x 81,071 ops/sec ±0.70% (63 runs sampled)
dyn 1ap  - fire first x 94,822 ops/sec ±0.83% (64 runs sampled)
dyn 5ap  - fire first x 47,884 ops/sec ±0.64% (64 runs sampled)
dyn 10ap - fire first x 29,547 ops/sec ±0.84% (64 runs sampled)
dyn 15ap - fire first x 21,282 ops/sec ±0.69% (63 runs sampled)
dyn 1ap  - fire last x 94,372 ops/sec ±0.82% (65 runs sampled)
dyn 5ap  - fire last x 46,844 ops/sec ±0.84% (65 runs sampled)
dyn 10ap - fire last x 28,860 ops/sec ±1.20% (63 runs sampled)
dyn 15ap - fire last x 21,060 ops/sec ±0.44% (64 runs sampled)
```