@zyla 2019-03-17
Seems like it isn't as good for performance as I thought.